Working on making the C agent support both windows and linux

+++ agent/C/src/striker_utils.c
@@ -143,7 +143,7 @@ void queue_free(queue *q, unsigned short items){
   free(q);
 }
 
-long find_offset(FILE *rfo, const void *target, size_t len){
+ssize_t find_offset(FILE *rfo, const void *target, size_t len){
 
   void *buff = malloc(len);
   int c;
